[fix] replace uint8_t flag to uint16_t flag

+++ os/lab4/mm/memory.c
@@ -32,7 +32,7 @@ unsigned char mem_map[PAGING_PAGES] = { 0 };
  *      - 地址必须按页对齐
  *      - 仅建立映射，不修改物理页引用计数
  */
-static inline void map_pages(uint64_t paddr_start, uint64_t paddr_end, uint64_t vaddr, uint8_t flag)
+static inline void map_pages(uint64_t paddr_start, uint64_t paddr_end, uint64_t vaddr, uint16_t flag)
 {
     while (paddr_start < paddr_end) {
         put_page(paddr_start, vaddr, flag);
@@ -164,7 +164,7 @@ uint64_t get_free_page(void)
  * @see panic(), map_kernel()
  * @note 地址需要按页对齐
  */
-uint64_t put_page(uint64_t page, uint64_t addr, uint8_t flag)
+uint64_t put_page(uint64_t page, uint64_t addr, uint16_t flag)
 {
     assert((page & (PAGE_SIZE - 1)) == 0,
            "put_page(): Try to put unaligned page %p to %p", page, addr);
@@ -192,7 +192,7 @@ uint64_t put_page(uint64_t page, uint64_t addr, uint8_t flag)
  * @param flag 标志位
  * @note 请确保该虚拟地址没有映射到物理页,否则本函数会覆盖原来的映射，导致内存泄漏。
  */
-void get_empty_page(uint64_t addr, uint8_t flag)
+void get_empty_page(uint64_t addr, uint16_t flag)
 {
     uint64_t tmp;
     assert(tmp = get_free_page(), "get_empty_page(): Memory exhausts");
